refactor: move locale related props to the abstract class
