+++ src/elements/displaynames/displaynames.ts
@@ -1,17 +1,9 @@
-import {PropertyValues} from 'lit';
 import {customElement, property} from 'lit/decorators.js';
 
 import AbstractIntlElement from '../abstract-intl-element';
-import {localesToLocaleList} from '../../utils/locales';
 
 @customElement('intl-displaynames')
 export default class DisplayNames extends AbstractIntlElement {
-  @property({attribute: false})
-  localeList: Intl.BCP47LanguageTag[] = [];
-
-  @property({reflect: true})
-  locales: Intl.LocalesArgument = '';
-
   @property({reflect: true})
   of = '';
 
@@ -21,21 +13,11 @@ export default class DisplayNames extends AbstractIntlElement {
   @property({attribute: 'intl-style', reflect: true})
   intlStyle: Intl.RelativeTimeFormatStyle = 'long';
 
-  @property({attribute: 'locale-matcher', reflect: true})
-  localeMatcher: Intl.RelativeTimeFormatLocaleMatcher = 'best fit';
-
   @property({attribute: 'language-display', reflect: true})
   languageDisplay: Intl.DisplayNamesLanguageDisplay = 'dialect';
 
   @property({reflect: true})
   fallback: Intl.DisplayNamesFallback = 'code';
-
-  protected override willUpdate(changes: PropertyValues<this>) {
-    if (changes.has('locales')) {
-      this.localeList =
-          localesToLocaleList(this.locales as string, this.localeMatcher);
-    }
-  }
 
   protected override render() {
     let result = '';
